Revert "Merge branch 'main' into East_zone"
This reverts commit 359a10ae0664cff736e40815f3e94a4195956413, reversing
changes made to 1a3d8b91ae627738a712f6c77d9f82fd02cb69bb.
