--- Assets/Prefabs/Menus.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f4a1d12b1272dea45837f938139cfab6
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
