--- Assets/Prefabs/Menus/PauseMenu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d99a2c65d771dd744927b206585782b1
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
