--- Assets/Prefabs/Menus/SimpleUpgradesMenu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bfbc8a2e9fbdbcd43b02039dc6fe32c6
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
