+++ Assets/Scripts/MenuController.cs
@@ -1,0 +1,38 @@
+using UnityEngine;
+
+public class MenuController : MonoBehaviour
+{
+    public static bool IsGamePaused = false;
+
+    public GameObject pauseMenuUI;
+
+    // Update is called once per frame
+    private void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.E))
+        {
+            if (IsGamePaused)
+            {
+                Resume();
+            } 
+            else
+            {
+                Pause();
+            }
+        }
+    }
+
+    private void Resume()
+    {
+        pauseMenuUI.SetActive(false);
+        Time.timeScale = 1f;
+        IsGamePaused = false;
+    }
+
+    private void Pause()
+    {
+        pauseMenuUI.SetActive(true);
+        Time.timeScale = 0f;
+        IsGamePaused = true;
+    }
+}
